+++ packages/localizations/src/en-US.ts
@@ -55,6 +55,7 @@ export const enUS: LocalizationResource = {
     seeAllFeatures: 'See all features',
     subscribe: 'Subscribe',
     switchPlan: 'Switch to this plan',
+    switchToMonthly: 'Switch to monthly',
     switchToAnnual: 'Switch to annual',
     viewFeatures: 'View features',
     year: 'Year',
@@ -152,7 +153,7 @@ export const enUS: LocalizationResource = {
     badge__unverified: 'Unverified',
     billingPage: {
       alerts: {
-        noPermissionsToManageBilling: 'You do not have permissions to manage billing for this account.',
+        noPermissionsToManageBilling: 'You do not have permissions to manage billing for this organization.',
       },
       paymentSourcesSection: {
         actionLabel__default: 'Make default',
@@ -180,6 +181,9 @@ export const enUS: LocalizationResource = {
       subscriptionsListSection: {
         actionLabel__switchPlan: 'Switch plans',
         title: 'Subscription',
+      },
+      switchPlansSection: {
+        title: 'Switch plans',
       },
       subscriptionsSection: {
         actionLabel__default: 'Manage',
@@ -777,6 +781,9 @@ export const enUS: LocalizationResource = {
       },
       subscriptionsSection: {
         actionLabel__default: 'Manage',
+      },
+      switchPlansSection: {
+        title: 'Switch plans',
       },
       title: 'Billing',
     },